+++ tests/test_plugins.py
@@ -611,6 +611,7 @@ class BadFileTracerTest(FileTracerTest):
         cov = coverage.Coverage()
         cov.set_option("run:plugins", [module_name])
         self.start_import_stop(cov, "simple")
+        cov.save()  # pytest-cov does a save after stop, so we'll do it too.
         return cov
 
     def run_bad_plugin(self, module_name, plugin_name, our_error=True, excmsg=None, excmsgs=None):
@@ -709,6 +710,37 @@ class BadFileTracerTest(FileTracerTest):
             class Plugin(coverage.plugin.CoveragePlugin):
                 def file_tracer(self, filename):
                     17/0 # Oh noes!
+
+            def coverage_init(reg, options):
+                reg.add_file_tracer(Plugin())
+            """)
+        self.run_bad_plugin("bad_plugin", "Plugin")
+
+    def test_file_tracer_fails_eventually(self):
+        # Django coverage plugin can report on a few files and then fail.
+        # https://github.com/nedbat/coveragepy/issues/1011
+        self.make_file("bad_plugin.py", """\
+            import os.path
+            import coverage.plugin
+            class Plugin(coverage.plugin.CoveragePlugin):
+                def __init__(self):
+                    self.calls = 0
+
+                def file_tracer(self, filename):
+                    print(filename)
+                    self.calls += 1
+                    if self.calls <= 2:
+                        return FileTracer(filename)
+                    else:
+                        17/0 # Oh noes!
+
+            class FileTracer(coverage.FileTracer):
+                def __init__(self, filename):
+                    self.filename = filename
+                def source_filename(self):
+                    return os.path.basename(self.filename).replace(".py", ".foo")
+                def line_number_range(self, frame):
+                    return -1, -1
 
             def coverage_init(reg, options):
                 reg.add_file_tracer(Plugin())